did some refactoring, and moved the categories and materials lists to the search model
