+++ lib/pages/searchPage/filterModalBottomsheet.dart
@@ -1,0 +1,58 @@
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+import 'package:nestify/models/searchModel.dart';
+import 'package:nestify/pages/searchPage/categoryFilterChips.dart';
+import 'package:nestify/pages/searchPage/materialFilterChips.dart';
+
+void filterModalBottomsheet(BuildContext context) {
+  showModalBottomSheet(
+    context: context,
+    isScrollControlled: true,
+    builder: (BuildContext context) {
+      return Container(
+        height: 450,
+        width: double.infinity,
+        padding: const EdgeInsets.all(24),
+        child: Column(
+            mainAxisAlignment: MainAxisAlignment.start,
+            crossAxisAlignment: CrossAxisAlignment.stretch,
+            children: [
+              Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                children: [
+                  Text('Filters',
+                      style: Theme.of(context).textTheme.titleLarge),
+                  TextButton(
+                      onPressed: () {
+                        var model =
+                            Provider.of<SearchModel>(context, listen: false);
+                        model.reset();
+                      },
+                      child: Text("Reset"))
+                ],
+              ),
+              const Padding(
+                padding: EdgeInsets.symmetric(vertical: 8),
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Padding(
+                      padding: EdgeInsets.only(bottom: 8),
+                      child: CategoryFilterChips(),
+                    ),
+                    MaterialFilterChips(),
+                  ],
+                ),
+              ),
+              Padding(
+                padding: const EdgeInsets.symmetric(vertical: 8),
+                child: FilledButton(
+                  onPressed: () {},
+                  child: const Text('Apply Filters'),
+                ),
+              ),
+            ]),
+      );
+    },
+  );
+}
